refactor(sortable): move from ui to blocks
Sortable comes from ReUI, not shadcn — relocate it to the blocks tree
so the ui/ directory continues to mirror shadcn 1:1.

Moves the source, example, and MDX files into the blocks layout, updates
the registry entry to `registry:block` (file type `registry:component`,
target `components/blocks/sortable.tsx`), and updates internal import
paths and MDX file/import references accordingly.
