+++ src/pages/blocks/kobalte/sortable.mdx
@@ -26,7 +26,7 @@ Install the following dependencies:
 
 Copy and paste the following code into your project.
 
-```tsx file=../../../registry/kobalte/ui/sortable.tsx showLineNumbers
+```tsx file=../../../registry/kobalte/blocks/sortable.tsx showLineNumbers
 
 ```
 
@@ -38,7 +38,7 @@ import {
   SortableItem,
   SortableItemHandle,
   SortableOverlay,
-} from "~/components/ui/sortable";
+} from "~/components/blocks/sortable";
 ```
 
 ```tsx
@@ -71,10 +71,10 @@ import { FileText, GripVertical, Image as ImageIcon, Music, Video } from "lucide
 import { createSignal, For } from "solid-js";
 import { toast } from "solid-sonner";
 import { Badge } from "~/components/ui/badge";
-import { Sortable, SortableItem, SortableItemHandle } from "~/components/ui/sortable";
+import { Sortable, SortableItem, SortableItemHandle } from "~/components/blocks/sortable";
 ```
 
-```tsx file=../../../registry/kobalte/examples/ui/sortable-example.tsx#L105-L151
+```tsx file=../../../registry/kobalte/examples/blocks/sortable-example.tsx#L105-L151
 
 ```
 
@@ -85,10 +85,10 @@ import { GripVertical } from "lucide-solid";
 import { createSignal, For } from "solid-js";
 import { toast } from "solid-sonner";
 import { Badge } from "~/components/ui/badge";
-import { Sortable, SortableItem, SortableItemHandle } from "~/components/ui/sortable";
+import { Sortable, SortableItem, SortableItemHandle } from "~/components/blocks/sortable";
 ```
 
-```tsx file=../../../registry/kobalte/examples/ui/sortable-example.tsx#L223-L267
+```tsx file=../../../registry/kobalte/examples/blocks/sortable-example.tsx#L223-L267
 
 ```
 
@@ -99,10 +99,10 @@ import { GripVertical } from "lucide-solid";
 import { createSignal, For } from "solid-js";
 import { toast } from "solid-sonner";
 import { Card, CardContent } from "~/components/ui/card";
-import { Sortable, SortableItem, SortableItemHandle } from "~/components/ui/sortable";
+import { Sortable, SortableItem, SortableItemHandle } from "~/components/blocks/sortable";
 ```
 
-```tsx file=../../../registry/kobalte/examples/ui/sortable-example.tsx#L311-L377
+```tsx file=../../../registry/kobalte/examples/blocks/sortable-example.tsx#L311-L377
 
 ```
 
@@ -118,11 +118,11 @@ import {
   CardHeader,
   CardTitle,
 } from "~/components/ui/card";
-import { Sortable, SortableItem, SortableItemHandle } from "~/components/ui/sortable";
+import { Sortable, SortableItem, SortableItemHandle } from "~/components/blocks/sortable";
 import { Switch } from "~/components/ui/switch";
 ```
 
-```tsx file=../../../registry/kobalte/examples/ui/sortable-example.tsx#L419-L463
+```tsx file=../../../registry/kobalte/examples/blocks/sortable-example.tsx#L419-L463
 
 ```
